Remove duplicate simulation script
Delete simulate.py in scripts, because it has exactly the same
functionality as sensitivity.py. Also rename sensitivity.py to
simulate.py in scripts and in tools and update all references in other
RMG files.
