+++ rmgpy/tools/data/sim/simple/species_dictionary.txt
