+++ scripts/simulate.py
@@ -29,14 +29,14 @@
 ###############################################################################
 
 """
-This script runs stand-alone simulation on an RMG job.  This is effectively the 
-same script as sensitivity.py
+This script runs a stand-alone simulation (including sensitivity analysis if
+specified in the input file) on an RMG job.
 """
 
 import os.path
 import argparse
 
-from rmgpy.tools.sensitivity import runSensitivity
+from rmgpy.tools.simulate import run_simulation
 
 ################################################################################
 
@@ -49,27 +49,23 @@ def parse_arguments():
         help='Chemkin file')
     parser.add_argument('dictionary', metavar='DICTIONARY', type=str, nargs=1,
         help='RMG dictionary file')
+    parser.add_argument('--no-dlim', dest='dlim', action='store_false',
+        help='Turn off diffusion-limited rates for LiquidReactor')
     args = parser.parse_args()
     
     inputFile = os.path.abspath(args.input[0])
     chemkinFile = os.path.abspath(args.chemkin[0])
     dictFile = os.path.abspath(args.dictionary[0])
+    dflag = args.dlim
 
-    return inputFile, chemkinFile, dictFile
+    return inputFile, chemkinFile, dictFile, dflag
 
 def main():
-    # This might not work anymore because functions were modified for use with webserver
+    inputFile, chemkinFile, dictFile, dflag = parse_arguments()
 
-    inputFile, chemkinFile, dictFile = parse_arguments()
-
-    runSensitivity(inputFile, chemkinFile, dictFile)
+    run_simulation(inputFile, chemkinFile, dictFile, dflag)
 
 ################################################################################
 
-
 if __name__ == '__main__':
     main()
-
-    
-    
-
